RSSBlogProcessor/requirements.txt
feedparser==6.0.10
pika==1.3.2
python-dotenv==1.0.0
sgmllib3k==1.0.0
